+++ lib/multiHtmlWebpackPlugin.js
@@ -5,70 +5,38 @@ const fs = require('fs');
 const _ = require('lodash');
 const path = require('path');
 const loaderUtils = require('loader-utils');
+const log = require('webpack-log');
 
-const { createHtmlTagObject, htmlTagObjectToString } = require('./lib/html-tags');
+const { htmlTagObjectToString } = require('./html-tags');
 
-const childCompiler = require('./lib/compiler.js');
-const prettyError = require('./lib/errors.js');
-const chunkSorter = require('./lib/chunksorter.js');
-const getHtmlWebpackPluginHooks = require('./lib/hooks.js').getHtmlWebpackPluginHooks;
-
+const childCompiler = require('./compiler.js');
+const prettyError = require('./errors.js');
+const chunkSorter = require('./chunksorter.js');
+const getHtmlWebpackPluginHooks = require('./hooks.js').getHtmlWebpackPluginHooks;
+const { setFileNameOption, setMinifyOption, setTemplateOption } = require('./optionsHelper.js');
 const fsStatAsync = promisify(fs.stat);
 const fsReadFileAsync = promisify(fs.readFile);
 
-class HtmlWebpackPluginUpdated {
+class multiHtmlWebpackPlugin {
     constructor(htmlWebpackPlugins, compiler) {
         this.htmlWebpackPlugins = htmlWebpackPlugins;
+
+        this.logger = log({ name: 'HtmlWebpackPlugin' })
 
         this.apply(compiler);
     }
 
     apply(compiler) {
+        this.logger.info('multiHtmlWebpackPlugin apply has been called')
         let isCompilationCached = false;
 
-        // Compile the template (queued)
-        const childCompilerPromises = [];
+        let childCompilerPromises;
 
-        this.htmlWebpackPlugins.forEach((plugin) => {
-            plugin.options.template = this.getFullTemplatePath(plugin.options.template, compiler.context);
-
-            // convert absolute filename into relative so that webpack can
-            // generate it at correct location
-            const filename = plugin.options.filename;
-            if (path.resolve(filename) === path.normalize(filename)) {
-                plugin.options.filename = path.relative(compiler.options.output.path, filename);
-            }
-
-            // `contenthash` is introduced in webpack v4.3
-            // which conflicts with the plugin's existing `contenthash` method,
-            // hence it is renamed to `templatehash` to avoid conflicts
-            plugin.options.filename = plugin.options.filename.replace(/\[(?:(\w+):)?contenthash(?::([a-z]+\d*))?(?::(\d+))?\]/ig, (match) => {
-                return match.replace('contenthash', 'templatehash');
-            });
-
-            // Check if webpack is running in production mode
-            // @see https://github.com/webpack/webpack/blob/3366421f1784c449f415cda5930a8e445086f688/lib/WebpackOptionsDefaulter.js#L12-L14
-            const isProductionLikeMode = compiler.options.mode === 'production' || !compiler.options.mode;
-
-            const minify = plugin.options.minify;
-            if (minify === true || (minify === 'auto' && isProductionLikeMode)) {
-                /** @type { import('html-minifier').Options } */
-                plugin.options.minify = {
-                    // https://github.com/kangax/html-minifier#options-quick-reference
-                    collapseWhitespace: true,
-                    removeComments: true,
-                    removeRedundantAttributes: true,
-                    removeScriptTypeAttributes: true,
-                    removeStyleLinkTypeAttributes: true,
-                    useShortDoctype: true
-                };
-            }
-        })
+        this.updateOptions(compiler);
 
         // Register all HtmlWebpackPlugins instances at the child compiler
         compiler.hooks.thisCompilation.tap('HtmlWebpackPlugin', (compilation) => {
-            debugger;
-            console.info('hook thisCompilation has been called')
+            this.logger.info('hook thisCompilation has been called')
             // Clear the cache if the child compiler is outdated
             if (childCompiler.hasOutDatedTemplateCache(compilation)) {
                 childCompiler.clearCache(compiler);
@@ -80,22 +48,19 @@ class HtmlWebpackPluginUpdated {
             // Add file dependencies of child compiler to parent compiler
             // to keep them watched even if we get the result from the cache
             compilation.hooks.additionalChunkAssets.tap('HtmlWebpackPlugin', () => {
-                debugger;
-                console.log('hook add additionalChunkAssets started')
+                this.logger.info('hook add additionalChunkAssets started')
                 const childCompilerDependencies = childCompiler.getFileDependencies(compiler);
                 childCompilerDependencies.forEach(fileDependency => {
                     compilation.compilationDependencies.add(fileDependency);
                 });
-                console.log('hook add additionalChunkAssets finished')
             });
-            console.info('hook thisCompilation has finished ')
         });
 
-        compiler.hooks.make.tapAsync('HtmlWebpackPlugin', (compilation, callback) => {
-            console.info('hook make has been called')
-            this.htmlWebpackPlugins.forEach((plugin) => {
-                debugger;
-                const childCompilerPromise = childCompiler.compileTemplate(plugin.options.template, plugin.options.filename, compilation)
+        compiler.hooks.make.tapPromise('HtmlWebpackPlugin', (compilation) => {
+            this.logger.info('hooks.make.tapPromise has been called')
+
+            childCompilerPromises = this.htmlWebpackPlugins.map((plugin) => {
+                return childCompiler.compileTemplate(plugin.options.template, plugin.options.filename, compilation)
                     .catch(err => {
                         compilation.errors.push(prettyError(err, compiler.context).toString());
                         return {
@@ -111,17 +76,9 @@ class HtmlWebpackPluginUpdated {
                         plugin.childCompilationOutputName = compilationResult.outputName;
                         return compilationResult.content;
                     });
-
-                childCompilerPromises.push(childCompilerPromise);
             });
 
-            Promise.all(childCompilerPromises)
-                .then((results) => {
-                    const a = results;
-                    console.info('hook make finished')
-                    debugger;
-                    callback();
-                })
+            return Promise.all(childCompilerPromises);
         });
 
         compiler.hooks.emit.tapAsync('HtmlWebpackPlugin',
@@ -131,7 +88,7 @@ class HtmlWebpackPluginUpdated {
              * @param {() => void} callback
             */
             (compilation, callback) => {
-                console.log('hook emit called');
+                this.logger.log('hook emit called');
                 let emitHtmlPromises = [];
 
                 this.htmlWebpackPlugins.forEach((plugin, index) => {
@@ -216,14 +173,12 @@ class HtmlWebpackPluginUpdated {
                             // Once everything is compiled evaluate the html factory
                             // and replace it with its content
 
-                            debugger;
                             return this.evaluateCompilationResult(compilation, compiledTemplate[index], plugin)
                         });
 
                     const templateExectutionPromise = Promise.all([assetsPromise, assetTagGroupsPromise, templateEvaluationPromise])
                         // Execute the template
                         .then(([assetsHookResult, assetTags, compilationResult]) => {
-                            debugger;
                             return typeof compilationResult !== 'function'
                                 ? compilationResult
                                 : this.executeTemplate(compilationResult, assetsHookResult.assets, { headTags: assetTags.headTags, bodyTags: assetTags.bodyTags }, compilation, plugin)
@@ -233,7 +188,6 @@ class HtmlWebpackPluginUpdated {
                     const injectedHtmlPromise = Promise.all([assetTagGroupsPromise, templateExectutionPromise])
                         // Allow plugins to change the html before assets are injected
                         .then(([assetTags, html]) => {
-                            debugger;
                             const pluginArgs = { html, headTags: assetTags.headTags, bodyTags: assetTags.bodyTags, plugin: plugin, outputName: childCompilationOutputName };
                             return getHtmlWebpackPluginHooks(compilation).afterTemplateExecution.promise(pluginArgs);
                         })
@@ -274,7 +228,7 @@ class HtmlWebpackPluginUpdated {
                             outputName: finalOutputName,
                             plugin: plugin
                         }).catch(err => {
-                            console.error(err);
+                            this.logger.error(err);
                             return null;
                         }).then(() => null));
 
@@ -285,10 +239,22 @@ class HtmlWebpackPluginUpdated {
                 // let the webpack compiler continue
                 Promise.all(emitHtmlPromises)
                     .then(() => {
-                        console.log('hook emit finished')
+                        this.logger.log('hook emit finished')
                         callback();
                     });
             });
+    }
+
+    updateOptions(compiler) {
+        this.logger.log('plugin options have been updated')
+        
+        this.htmlWebpackPlugins.forEach((htmlWebpackPlugin) => {
+            htmlWebpackPlugin.options.template = setTemplateOption(htmlWebpackPlugin.options.template, compiler.context);
+
+            htmlWebpackPlugin.options.filename = setFileNameOption(htmlWebpackPlugin.options, compiler);
+
+            htmlWebpackPlugin.options.minify = setMinifyOption(htmlWebpackPlugin.options, compiler);
+        })
     }
 
     /**
@@ -298,7 +264,6 @@ class HtmlWebpackPluginUpdated {
    * @returns {Promise<string | (() => string | Promise<string>)>}
    */
     evaluateCompilationResult(compilation, source, plugin) {
-        debugger;
         if (!source) {
             return Promise.reject(new Error('The child compilation didn\'t provide a result'));
         }
@@ -318,7 +283,7 @@ class HtmlWebpackPluginUpdated {
         if (typeof newSource === 'object' && newSource.__esModule && newSource.default) {
             newSource = newSource.default;
         }
-        debugger;
+
         return typeof newSource === 'string' || typeof newSource === 'function'
             ? Promise.resolve(newSource)
             : Promise.reject(new Error('The loader "' + plugin.options.template + '" didn\'t return html.'));
@@ -380,7 +345,6 @@ class HtmlWebpackPluginUpdated {
         const templateParamsPromise = this.getTemplateParameters(compilation, assets, assetTags, plugin);
         return templateParamsPromise.then((templateParams) => {
             try {
-                debugger;
                 // If html is a promise return the promise
                 // If html is a string turn it into a promise
                 return templateFunction(templateParams);
@@ -406,7 +370,6 @@ class HtmlWebpackPluginUpdated {
      * @returns {Promise<string>}
      */
     postProcessHtml(html, assets, assetTags, plugin) {
-        debugger;
         if (typeof html !== 'string') {
             return Promise.reject(new Error('Expected html to be a string but got ' + JSON.stringify(html)));
         }
@@ -877,30 +840,6 @@ class HtmlWebpackPluginUpdated {
     }
 
     /**
-     * Helper to return the absolute template path with a fallback loader
-     * @param {string} template
-     * The path to the template e.g. './index.html'
-     * @param {string} context
-     * The webpack base resolution path for relative paths e.g. process.cwd()
-     */
-    getFullTemplatePath(template, context) {
-        if (template === 'auto') {
-            template = path.resolve(context, 'src/index.ejs');
-            if (!fs.existsSync(template)) {
-                template = path.join(__dirname, 'default_index.ejs');
-            }
-        }
-        // If the template doesn't use a loader use the lodash template loader
-        if (template.indexOf('!') === -1) {
-            template = require.resolve('./lib/loader.js') + '!' + path.resolve(context, template);
-        }
-        // Resolve template path
-        return template.replace(
-            /([!])([^/\\][^!?]+|[^/\\!?])($|\?[^!?\n]+$)/,
-            (match, prefix, filepath, postfix) => prefix + path.resolve(filepath) + postfix);
-    }
-
-    /**
      * Helper to return a sorted unique array of all asset files out of the
      * asset object
      */
@@ -911,4 +850,6 @@ class HtmlWebpackPluginUpdated {
     }
 }
 
-module.exports = HtmlWebpackPluginUpdated;
+module.exports = {
+    multiHtmlWebpackPlugin
+};
